cpufreq: Fix broken uevents for cpufreq governor and cpu devices
cyanogens uevent commit was rendered non-working since kitkat (or so)
as the uevent filter function caused events to be dropped. Create an
empty filter to cure this. Then hook into cpufreq_core_init() and
cpufreq_add_dev_interface() to create our ksets for cpufreq and cpu
devices.

This commit requires a change in ueventd.rc that add rules for several
files of interest.

Signed-off-by: tarun93 <tarunmyid@gmail.com>